Updating README.md
I'm always hesitant to write documentation as it so often has lied to
me. But there is something comforting about seeing that there a gem
has documentation.

It shows that at one point, someone agreed to spend a bit of time
saying in English how they believe the code works. But is that really
how it works? I'd rather rely on tests to verify that behavior.

But I digress. I'm opting to instead document what I believe to be the
external API of Mappy. It is the least likely to change and be
inaccurate. It is also the point in which a bit of English can help
explain the code.